COMP: Fix VS compiler warnings (#855)

+++ src/Plugins/SimplnxCore/test/ReadHDF5DatasetTest.cpp
@@ -53,7 +53,7 @@ void writePointerArrayDataset(nx::core::HDF5::GroupWriter& ptrGroupWriter)
     dims = {10, 8, 6, (COMPDIMPROD * TUPLEDIMPROD) / 10 / 8 / 6};
   }
 
-  hsize_t tSize = std::accumulate(dims.cbegin(), dims.cend(), 1UL, std::multiplies<hsize_t>());
+  hsize_t tSize = std::accumulate(dims.cbegin(), dims.cend(), static_cast<usize>(1), std::multiplies<hsize_t>());
   std::vector<T> data(tSize);
   for(hsize_t i = 0; i < tSize; ++i)
   {
